Move observability to pkg.
