+++ pkg/cloudevents/datacodec/observability.go
@@ -1,7 +1,7 @@
 package datacodec
 
 import (
-	"github.com/cloudevents/sdk-go/pkg/cloudevents/observability"
+	"github.com/cloudevents/sdk-go/pkg/observability"
 	"go.opencensus.io/stats"
 	"go.opencensus.io/stats/view"
 )
