+++ pkg/cloudevents/transport/http/observability.go
@@ -3,7 +3,7 @@ package http
 import (
 	"fmt"
 
-	"github.com/cloudevents/sdk-go/pkg/cloudevents/observability"
+	"github.com/cloudevents/sdk-go/pkg/observability"
 	"go.opencensus.io/stats"
 	"go.opencensus.io/stats/view"
 )
